% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblefiddle.r
\name{tibblefiddle}
\alias{tibblefiddle}
\title{Fiddle with  a \code{tibble} Object}
\usage{
tibblefiddle(x, annotate_vars = names(x), hide_vars = character(), ...)
}
\arguments{
\item{x}{the dataframe that will be annotated.}

\item{annotate_vars}{the names of columns to annotate.}

\item{hide_vars}{the columns to hide in the shiny app.}

\item{...}{currently unused.}
}
\value{
the modified tibble
}
\description{
Modify the values of specified columns in a dataframe in shiny.
}
