% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbletate.r
\name{tibbletate}
\alias{tibbletate}
\title{Annotate a \code{tibble} Ojbect}
\usage{
tibbletate(x, annotate_cols, ...)
}
\arguments{
\item{x}{the dataframe that will be annotated.}

\item{annotate_cols}{the names of columns to annotate.}

\item{...}{currently unused.}
}
\value{
the modified tibble
}
\description{
Modify the values of specified columns in a dataframe in shiny.
}
